MapExplorer split to MapVisuals
